combine props and include test coverage for getRuleValue func
